requirements.txt
numpy==2.4.4
pandas==3.0.2
PyYAML==6.0.3
